Minor refactoring, reduce precision in many places where not needed

+++ OsuDiffCalc/FileProcessor/BeatmapObjects/BreakSection.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace OsuDiffCalc.FileProcessor.BeatmapObjects {
 	class BreakSection : BeatmapObject {
-		public BreakSection(double startTime, double endTime) : base(startTime, endTime) {
+		public BreakSection(int startTime, int endTime) : base(startTime, endTime) {
 		}
 	}
 }
