+++ OsuDiffCalc/FileProcessor/BeatmapObjects/Spinner.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace OsuDiffCalc.FileProcessor.BeatmapObjects {
 	class Spinner : HitObject {
-		public Spinner(float x, float y, double startTime, double endTime) : base(x, y, startTime, endTime) {
+		public Spinner(float x, float y, int startTime, int endTime) : base(x, y, startTime, endTime) {
 		}
 	}
 }
